@@ -1,62 +1,164 @@
 {
-  "name": "libp2p-record",
+  "name": "@libp2p/record",
   "version": "0.10.6",
   "description": "libp2p record implementation",
-  "leadMaintainer": "Vasco Santos <vasco.santos@moxy.studio>",
-  "main": "src/index.js",
-  "scripts": {
-    "test": "aegir test",
-    "lint": "aegir lint",
-    "test:node": "aegir test -t node",
-    "test:browser": "aegir test -t browser -t webworker",
-    "prepare": "npm run build",
-    "build": "run-s build:*",
-    "build:types": "aegir build --no-bundle",
-    "build:proto": "pbjs -t static-module -w commonjs -r libp2p-record --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/record/record.js ./src/record/record.proto",
-    "build:proto-types": "pbts -o src/record/record.d.ts src/record/record.js",
-    "docs": "aegir docs",
-    "release": "aegir release",
-    "release-minor": "aegir release --type minor",
-    "release-major": "aegir release --type major",
-    "coverage": "aegir coverage"
-  },
+  "author": "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/libp2p/js-libp2p-record#readme",
   "repository": {
     "type": "git",
-    "url": "https://github.com/libp2p/js-libp2p-record.git"
+    "url": "git+https://github.com/libp2p/js-libp2p-record.git"
+  },
+  "bugs": {
+    "url": "https://github.com/libp2p/js-libp2p-record/issues"
   },
   "keywords": [
     "IPFS"
   ],
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
   },
-  "pre-push": [
-    "lint"
-  ],
-  "author": "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/libp2p/js-libp2p-record/issues"
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "*": [
+        "*",
+        "dist/*",
+        "dist/src/*",
+        "dist/src/*/index"
+      ],
+      "src/*": [
+        "*",
+        "dist/*",
+        "dist/src/*",
+        "dist/src/*/index"
+      ]
+    }
   },
-  "homepage": "https://github.com/libp2p/js-libp2p-record",
   "files": [
     "src",
-    "dist"
+    "dist/src",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
   ],
+  "exports": {
+    ".": {
+      "import": "./dist/src/index.js"
+    },
+    "./selectors": {
+      "import": "./dist/src/selectors.js"
+    },
+    "./validators": {
+      "import": "./dist/src/validators.js"
+    }
+  },
   "eslintConfig": {
     "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    },
     "ignorePatterns": [
-      "src/record/record.d.ts"
+      "src/record.d.ts"
     ]
   },
-  "types": "dist/src/index.d.ts",
-  "devDependencies": {
-    "aegir": "^35.0.1",
-    "libp2p-crypto": "^0.19.6",
-    "libp2p-interfaces": "^1.0.1",
-    "npm-run-all": "^4.1.5",
-    "peer-id": "^0.15.2",
-    "util": "^0.12.4"
+  "release": {
+    "branches": [
+      "master"
+    ],
+    "plugins": [
+      [
+        "@semantic-release/commit-analyzer",
+        {
+          "preset": "conventionalcommits",
+          "releaseRules": [
+            {
+              "breaking": true,
+              "release": "major"
+            },
+            {
+              "revert": true,
+              "release": "patch"
+            },
+            {
+              "type": "feat",
+              "release": "minor"
+            },
+            {
+              "type": "fix",
+              "release": "patch"
+            },
+            {
+              "type": "chore",
+              "release": "patch"
+            },
+            {
+              "type": "docs",
+              "release": "patch"
+            },
+            {
+              "type": "test",
+              "release": "patch"
+            },
+            {
+              "scope": "no-release",
+              "release": false
+            }
+          ]
+        }
+      ],
+      [
+        "@semantic-release/release-notes-generator",
+        {
+          "preset": "conventionalcommits",
+          "presetConfig": {
+            "types": [
+              {
+                "type": "feat",
+                "section": "Features"
+              },
+              {
+                "type": "fix",
+                "section": "Bug Fixes"
+              },
+              {
+                "type": "chore",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "docs",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "test",
+                "section": "Tests"
+              }
+            ]
+          }
+        }
+      ],
+      "@semantic-release/changelog",
+      "@semantic-release/npm",
+      "@semantic-release/github",
+      "@semantic-release/git"
+    ]
+  },
+  "scripts": {
+    "lint": "aegir lint",
+    "pretest": "npm run build",
+    "test": "aegir test -f ./dist/test",
+    "test:node": "npm run test -- -t node",
+    "test:chrome": "npm run test -- -t browser",
+    "test:chrome-webworker": "npm run test -- -t webworker",
+    "test:firefox": "npm run test -- -t browser -- --browser firefox",
+    "test:firefox-webworker": "npm run test -- -t webworker -- --browser firefox",
+    "build": "tsc",
+    "postbuild": "npm run build:copy-proto-files",
+    "generate:proto": "pbjs -t static-module -w es6 -r libp2p-record --force-number --no-verify --no-delimited --no-create --no-beautify --no-defaults --lint eslint-disable -o src/record/record.js ./src/record/record.proto",
+    "generate:proto-types": "pbts -o src/record/record.d.ts src/record/record.js",
+    "build:copy-proto-files": "cp src/record* dist/src",
+    "release": "semantic-release"
   },
   "dependencies": {
     "err-code": "^3.0.1",
