@@ -1,18 +1,14 @@
 /* eslint-env mocha */
-'use strict'
-
-const { expect } = require('aegir/utils/chai')
-const { fromString: uint8ArrayFromString } = require('uint8arrays/from-string')
-const libp2pRecord = require('../src')
-const Record = libp2pRecord.Record
-
-const fixture = require('./fixtures/go-record.js')
+import { expect } from 'aegir/utils/chai.js'
+import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
+import { Libp2pRecord } from '../src/index.js'
+import * as fixture from './fixtures/go-record.js'
 
 const date = new Date(Date.UTC(2012, 1, 25, 10, 10, 10, 10))
 
 describe('record', () => {
   it('new', () => {
-    const rec = new Record(
+    const rec = new Libp2pRecord(
       uint8ArrayFromString('hello'),
       uint8ArrayFromString('world')
     )
@@ -22,8 +18,8 @@ describe('record', () => {
   })
 
   it('serialize & deserialize', () => {
-    const rec = new Record(uint8ArrayFromString('hello'), uint8ArrayFromString('world'), date)
-    const dec = Record.deserialize(rec.serialize())
+    const rec = new Libp2pRecord(uint8ArrayFromString('hello'), uint8ArrayFromString('world'), date)
+    const dec = Libp2pRecord.deserialize(rec.serialize())
 
     expect(dec).to.have.property('key').eql(uint8ArrayFromString('hello'))
     expect(dec).to.have.property('value').eql(uint8ArrayFromString('world'))
@@ -32,7 +28,7 @@ describe('record', () => {
 
   describe('go interop', () => {
     it('no signature', () => {
-      const dec = Record.deserialize(fixture.serialized)
+      const dec = Libp2pRecord.deserialize(fixture.serialized)
       expect(dec).to.have.property('key').eql(uint8ArrayFromString('hello'))
       expect(dec).to.have.property('value').eql(uint8ArrayFromString('world'))
     })
